ail: Add compressed miptree offset & tilesize tests
An even bigger pile of tests... This adds general miptree tests for some
compressed formats, and even more comprehensive miptree and tile size
tests for more formats and size combinations.

Since the tests are based on observing tiling done by the Metal driver,
we don't know the actual tile size, but rather we can just identify
which tile sizes logically have the same result (since several sizes can
be equivalent). This is encoded as a bit mask, split into two halves to
handle the case of stride padding for compressed textures (which tile
sizes are valid with and without stride padding can vary, and sometimes
you can either change the tile size or add padding and end up with the
same result). As long as whatever configuration the layout code comes up
with has its corresponding bit set in the bit mask, the tiling should be
correct.

Signed-off-by: Asahi Lina <lina@asahilina.net>