agx: fix RA shuffling algorithm
to guarantee that we find a shuffle, we need to shuffle in descending order of
alignment. this ensures we don't refragment the part of the register file we're
attempting to defragment. rewrite the shuffle code accodingly, using a worklist
that picks a largest variable to shuffle at each iteration.

fixes crashes with =spill.

Signed-off-by: Alyssa Rosenzweig <alyssa@rosenzweig.io>